[M1_TR-212] Moved components into folders
